bookingservice/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
